+++ tests/unit/test-dot.cpp
@@ -13,7 +13,7 @@ int main(int argc, char ** argv)
     std::array< Real, dim > u = { Real(1.0), Real(2.0), Real(3.0) };
     std::array< Real, dim > v = { Real(4.0), Real(5.0), Real(6.0) };
     const Real dot_product = Dot( u, v );
-    std::cout << "Dot product: " << dot_product << std::endl;
+    std::cout << "Dot product: " << double(dot_product) << std::endl;
     if (dot_product == Real(32.0))
     {
         return 0;
